module/ai/documentai/python/requirements.txt
google-cloud-storage
google-cloud-documentai